requirements.txt
pydantic>=2.6
orjson>=3.10
pyyaml>=6.0
tqdm>=4.66
pandas>=2.2
numpy>=1.26
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
rapidfuzz>=3.6
PyMuPDF>=1.24
